refactor : 전체적으로 코드 수정 - product, promotion exception 클래스 분리 - promotion_products 테이블에 updated_at 컬럼추가로 인한 data.sql,  schema.sql 수정 - 사용하지 않는 errorcode 제거
