+++ src/test/java/antigravity/service/ProductServiceExceptionTest.java
@@ -8,7 +8,6 @@ import static org.mockito.BDDMockito.then;
 
 import antigravity.exception.ProductApplicationException;
 import antigravity.exception.code.ProductErrorCode;
-import antigravity.exception.code.PromotionErrorCode;
 import antigravity.fixture.ProductFixture;
 import antigravity.fixture.PromotionProductFixture;
 import antigravity.model.request.ProductInfoRequest;
@@ -23,14 +22,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 
-
 @SpringBootTest
-class CalculateProductPriceExceptionTest {
+public class ProductServiceExceptionTest {
 
     @Autowired private CalculateProductPriceService calculateProductPriceService;
     @MockBean private ProductRepository productRepository;
     @MockBean private PromotionProductsRepository promotionProductsRepository;
-
 
     @Test
     @DisplayName("해당 상품이 존재하지 않는다면 ProductApplicationException(NOT_EXIST_PRODUCT)을 발생시킨다.")
@@ -43,48 +40,6 @@ class CalculateProductPriceExceptionTest {
 
         assertEquals(ProductErrorCode.NOT_EXIST_PRODUCT, e.getErrorCode());
         then(productRepository).should().findById(any(Integer.class));
-    }
-
-    @Test
-    @DisplayName("중복되는 프로모션 적용이 요청된다면 ProductApplicationException(DUPLICATED_PROMOTION)을 발생시킨다.")
-    void promotionDuplicationException() {
-        ProductInfoRequest pir = ProductInfoRequest.builder().productId(1).couponIds(new int[]{1, 1}).build();
-        given(productRepository.findById(1)).willReturn(Optional.ofNullable(ProductFixture.getProduct()));
-
-        ProductApplicationException e = assertThrows(ProductApplicationException.class,
-            () -> calculateProductPriceService.getProductAmount(pir));
-
-        assertEquals(PromotionErrorCode.DUPLICATED_PROMOTION, e.getErrorCode());
-        then(productRepository).should().findById(any(Integer.class));
-    }
-
-    @Test
-    @DisplayName("해당 프로모션의 기한이 지났다면 ProductApplicationException(INVALID_PROMOTION_PERIOD)을 발생시킨다.")
-    void promotionPeriodInvalidException() {
-        ProductInfoRequest pir = ProductInfoRequest.builder().productId(1).couponIds(new int[]{1, 2}).build();
-        given(productRepository.findById(1)).willReturn(Optional.ofNullable(ProductFixture.getProduct()));
-        given(promotionProductsRepository.findWithPromotionByPromotionIdIn(Arrays.asList(1, 2))).willReturn(
-            Collections.singletonList(PromotionProductFixture.getExpiredPeriodPromotionProducts()));
-
-        ProductApplicationException e = assertThrows(ProductApplicationException.class,
-            () -> calculateProductPriceService.getProductAmount(pir));
-
-        assertEquals(PromotionErrorCode.INVALID_PROMOTION_PERIOD, e.getErrorCode());
-    }
-
-    @Test
-    @DisplayName("해당 프로모션의에 해당되는 상품이 아니라면 ProductApplicationException(INVALID_PROMOTION_PRODUCT)을 발생시킨다.")
-    void promotionOfProductInvalidException() {
-        ProductInfoRequest pir = ProductInfoRequest.builder().productId(1).couponIds(new int[]{1, 2}).build();
-
-        given(productRepository.findById(1)).willReturn(Optional.ofNullable(ProductFixture.getProduct()));
-        given(promotionProductsRepository.findWithPromotionByPromotionIdIn(Arrays.asList(1, 2))).willReturn(
-            Collections.singletonList(PromotionProductFixture.getInValidPromotionProducts()));
-
-        ProductApplicationException e = assertThrows(ProductApplicationException.class,
-            () -> calculateProductPriceService.getProductAmount(pir));
-
-        assertEquals(PromotionErrorCode.INVALID_PROMOTION_PRODUCT, e.getErrorCode());
     }
 
     @Test
@@ -115,21 +70,6 @@ class CalculateProductPriceExceptionTest {
             () -> calculateProductPriceService.getProductAmount(pir));
 
         assertEquals(ProductErrorCode.MAX_PRICE_LIMIT, e.getErrorCode());
-    }
-
-    @Test
-    @DisplayName("해당 프로모션 할인가격이 기존 제품 가격보다 초과된다면 ProductApplicationException(EXCEED_ORIGIN_PRICE)을 발생시킨다.")
-    void promotionPriceExceedProductPriceException() {
-        ProductInfoRequest pir = ProductInfoRequest.builder().productId(1).couponIds(new int[]{1, 2}).build();
-
-        given(productRepository.findById(1)).willReturn(Optional.ofNullable(ProductFixture.getProduct()));
-        given(promotionProductsRepository.findWithPromotionByPromotionIdIn(Arrays.asList(1, 2))).willReturn(
-            Collections.singletonList(PromotionProductFixture.getExceedPromotionPriceProducts()));
-
-        ProductApplicationException e = assertThrows(ProductApplicationException.class,
-            () -> calculateProductPriceService.getProductAmount(pir));
-
-        assertEquals(PromotionErrorCode.EXCEED_ORIGIN_PRICE, e.getErrorCode());
     }
 
 }
